created transition log component
