+++ tests/dummy/app/components/logged-between-components.js
@@ -1,0 +1,5 @@
+import BetweenExample from './between-components';
+import layout from '../templates/components/between-components';
+import { extensions } from './transition-log-table';
+
+export default BetweenExample.extend({ layout }).extend(extensions);
